Sync with upstream (#1)
* Playlists: Auto update on changes/ Add ability to delete

A couple other notable changes:

- <Icon> size prop now works correctly
- The "Apple Music" red color is now correct

* Playlist View: Add red "Options" circle

This is just a style change. The functionality is exactly the same
(When the button is clicked, the playlist options pops up)

* Update README.md

* Update README.md

* Create LICENSE

* Dependencies: Update

There are a couple vulnerabilities that are fixed in this commit.

* Album Images: Lazy load & update placeholder

Should've done this a while ago, but images now lazy load and include a
placeholder image while the image loads. Once it's loaded, the
placeholder will fade out.

* View Management: Clean up code

It's been a while since I've messed with how views are displayed, and
since that time I've learned a few new things on how to organize views.
The code for determining the state of a view was pretty confusing. While
it's still a little convoluted, I think I've managed to get everything
looking a little better. I'd like to move `<ViewStack>` out of
`view_container.js` soon, but I'll save that for another time.